area of interests with faculty names/requirements.txt
Flask
selenium
webdriver-manager
mysql-connector-python
